fix sample names in dataframe
Before this commit, the file included the absolute path, which was 
included in the dataframe as "Sample_name". Only the extension was
removed. Now also the file path is removed by splitting on slashes and
taking the final element. (Changing sample names from e.g.
'/path/to/my_results/sample1.extension' to 'sample1'.)